Move ai integration tests into expected-use
